+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     Fix issue with `kapacitord config > kapacitor.conf` when the output file was a default location for the config.
 - [#626](https://github.com/influxdata/kapacitor/issue/626): Fix issues when changing the ID of an enabled task.
 - [#624](https://github.com/influxdata/kapacitor/pull/624): Fix issues where you could get a read error on a closed UDF socket.
+- [#651](https://github.com/influxdata/kapacitor/pull/651): Fix issues where an error during a batch replay would hang because the task wouldn't stop.
 
 
 ## v1.0.0-beta1 [2016-06-06]